+++ core/collection.ts
@@ -74,15 +74,6 @@ export default class Collection {
       delete: options.delete ?? true,
       rename: options.rename ?? true,
     };
-
-    if (
-      this.permissions.rename !== true && this.permissions.create &&
-      !this.documentName
-    ) {
-      throw new Error(
-        "The documentName option is required if the rename permission is false or auto.",
-      );
-    }
   }
 
   get fields() {